+++ building-service/src/main/java/com/islandempires/buildingservice/shared/buildinglevelspec/IslandHeadquarterLevel.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class IslandHeadquarterLevel extends BuildingLevel {
 
-    private double timeReductionPercentage;
+    private int timeReductionPercentage;
 
 }
