+++ gameserver-service/src/main/java/com/islandempires/gameserverservice/model/buildinglevelspec/WatchTowerLevel.java
@@ -4,12 +4,14 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class WatchTowerLevel extends BuildingLevel {
 
-    private Double visibilityRange;
+    private int visibilityRange;
 
-    private Integer observableCapacity;
+    private BigInteger observableCapacity;
 }
